requirements.txt
streamlit
python-dotenv
langdetect
pdf2image
llama-parse
openai
PyMuPDF
python-docx
Pillow
